requirements.txt
annotated-types==0.5.0
anyio==3.7.1
APScheduler==3.6.3
bandcamp-api==0.1.15
beautifulsoup4==4.12.2
Brotli==1.0.9
bs4==0.0.1
cachetools==4.2.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
dacite==1.8.1
demjson3==3.0.6
Deprecated==1.2.14
fastapi==0.100.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httpx==0.27.0
idna==3.4
jaconv==0.3.4
lxml==4.9.3
markdown-it-py==3.0.0
mdurl==0.1.2
mutagen==1.46.0
platformdirs==3.10.0
pycryptodomex==3.18.0
pydantic==2.1.1
pydantic_core==2.4.0
Pygments==2.16.1
pykakasi==2.2.1
python-dateutil==2.8.2
python-slugify==8.0.1
python-telegram-bot==13.11
pytube==15.0.0
pytz==2024.1
rapidfuzz==2.15.1
redis==4.6.0
requests==2.31.0
rich==13.5.2
six==1.16.0
sniffio==1.3.0
soundcloud-v2==1.3.1
soupsieve==2.4.1
spotdl==4.2.0
spotipy==2.23.0
starlette==0.27.0
syncedlyrics==0.5.0
text-unidecode==1.3
tornado==6.4
typing_extensions==4.7.1
tzdata==2024.1
tzlocal==5.2
Unidecode==1.3.6
urllib3==2.0.4
uvicorn==0.22.0
webencodings==0.5.1
websockets==11.0.3
wrapt==1.15.0
yt-dlp==2023.7.6
ytmusicapi==0.24.1
